pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "authlib>=1.6.1",
    "discord-py>=2.5.2",
    "flask-session>=0.8.0",
    "flask>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.42",
    "requests>=2.32.4",
]
